requirements.txt
Flask==3.0.0
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.28
gspread==5.12.0
oauth2client==4.1.3
PyYAML==6.0.1
Werkzeug==3.0.1
Jinja2==3.1.2 